Implement user specified x/y_val datasets
To allow for a user to potentially augment their training data but
not their validation data (and it should be noted that nobody should
ever augment first then randomize then split, else major bias
problems), implement an option to explicitly specify the validation
dataset.

Also make necessary changes to validation_split to account for this,
and make a few QOL edits to the testing suite.