major namespace cleanup and experimental automl tools
1) experimental automl capabilities:

- added Params() for generating parameter dictionaries
- added KerasModel() for generating network architectures

2) namespace cleaning:

The namespace now consist only of actionable/useful items. The top-level consist of commands which all are classes so can be identified being camel-case and utils and templates lead to second-level items. Check it out for yourself to learn more. Further, examples has a third-level which is datasets, models, and pipelines.

4) fixes and cleanups

- cleaned up the naming in ParamGrid
- changed all the numpy imports to come from np
- improved scan_object.data to numeric conversion

4) added new tests