testing restructure and permutation_filter
- moved permutation_filter to its own function. The question here is that would it be meaningful to support boolean statement directly, or is there benefit from having the lambda?
- premutation >> permutation
- fixed the case where unrecognized random_method failed silently, no warns the user and falls back to 'uniform_mersenne'
- added missing doctrings to Scan()
- the tests are now organized in a meaningful way:

"""
The tests below have to serve several purpose:

- test possible input methods to params dict
- test binary, multi class, multi label and continuous problems
- test all Scan arguments

Each problem type is presented as a Class, and contains three
experiments using single, list, or range inputs. There is an
effort to test as many scenarios as possible here, so be
inventive / experiment! Doing well with this part of the testing,
there is a healthy base for a more serious approach to ensuring
procedural integrity.

"""