Major overhaul of the codebase + ParamSpace API
This version will introduce a self-contained ParamSpace API allowing Scan() a single line interface to qualified permutations. The change allows important streamlining of the main procedural codes of Talos. These simplifications are already reflected in this commit.

Other changes:

- a major rehaul of reducers; reduction strategies can now be easily added in a single file/function
- it's now very
- moved all logging/results related codes to /logging
- Deals with #153
- Major cleaning up of Scan() arguments
- Adds a reducer that takes in a metric, threshold, and loss e.g ['val_acc', 0.9, False] where the experiment will be ended once a given metric threshold is met by a model
- /examples is now /templates
- many redundant functions / files were deleted
- ~100 lines of code were removed from the mainline codes and those codes were notably streamlined
- reduction no longer has prepare or finish
- learning entropy (metrics/entropy.py) is completely rewritten

A big bunch of other things, so do check it out for yourself.

NOTE: this version is still under testing.