Fix of parameter space
Shouldn't there be brackets in the parameter space for dropout and the learning rate? In my runs it's just selecting the first value. Also put some values in that seemed more reasonable. 

It actually does some pretty weird things if you put parentheses in for some of the other parameter options as well. first_neuron and batch_size will run the space for more options than would be in the multiplicative space. In the others I checked it thankfully throws an error though.

Thanks for this!