v.0.6.2 docs and feature update
The most import aspect of this version update is the new docs. The docs are comprehensive, and consistent with the >0.6 Talos API.

Some small updates and fixes outside of the docs are included in this update.

- Reporting() now also accepts Analyze() as its command
- Reporting() no longer has 'val_acc' as default value in any of the properties
- early_stopper() now has 'lazy' mode with slight tweaks to other modes
- early_stopper() no longer expects epochs value for custom settings
- Scan() docstrings are up-to-date and cleaned up
- some other docstring cleanups are present
- Docstrings will take the same format in the future as markup table to give consistency with documentations

NOTE: the docs are a work in progress and will be completed throughout this week. Pushing now to avoid the case where for some reason several days of work will be lost.