"NO DEPLOY FRIDAYS" and major cleanup of tests
...is more than just a catchy phrase.

In any case, the test suite is now clean and extensible as opposed to the coverage obsessed mess that it was before.

Quite a few bugs were squashed betwee now and Friday 4pm. Most importantly, there were quite a few sloppiness in the reducer process, that would have effected the occassional case.